+++ src/main/java/com/nexters/goalpanzi/application/mission/MissionMemberService.java
@@ -4,13 +4,12 @@ import com.nexters.goalpanzi.application.firebase.TopicGenerator;
 import com.nexters.goalpanzi.application.mission.dto.response.MemberRankResponse;
 import com.nexters.goalpanzi.application.mission.dto.response.MissionDetailResponse;
 import com.nexters.goalpanzi.application.mission.dto.response.MissionsResponse;
-import com.nexters.goalpanzi.application.mission.event.*;
+import com.nexters.goalpanzi.application.mission.event.CancelMissionRetryPushMessageEvent;
+import com.nexters.goalpanzi.application.mission.event.JoinMissionEvent;
+import com.nexters.goalpanzi.application.mission.event.SubscribeToMissionEvent;
 import com.nexters.goalpanzi.domain.common.BaseEntity;
-import com.nexters.goalpanzi.domain.device.Device;
-import com.nexters.goalpanzi.domain.device.DeviceSubscription;
 import com.nexters.goalpanzi.domain.device.Devices;
 import com.nexters.goalpanzi.domain.device.repository.DeviceRepository;
-import com.nexters.goalpanzi.domain.device.repository.DeviceSubscriptionRepository;
 import com.nexters.goalpanzi.domain.member.Member;
 import com.nexters.goalpanzi.domain.member.repository.MemberRepository;
 import com.nexters.goalpanzi.domain.mission.*;
@@ -20,7 +19,6 @@ import com.nexters.goalpanzi.exception.AlreadyExistsException;
 import com.nexters.goalpanzi.exception.ErrorCode;
 import com.nexters.goalpanzi.exception.NotFoundException;
 import com.nexters.goalpanzi.infrastructure.firebase.PushMessageSender;
-import com.nexters.goalpanzi.infrastructure.firebase.TopicSubscriber;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.stereotype.Service;
@@ -28,11 +26,9 @@ import org.springframework.transaction.annotation.Transactional;
 
 import java.time.LocalDateTime;
 import java.util.List;
-import java.util.stream.Collectors;
 
 import static com.nexters.goalpanzi.domain.firebase.PushMessage.MISSION_CANCELLATION_WARNING;
 import static com.nexters.goalpanzi.domain.firebase.PushMessage.MISSION_READY;
-import static com.nexters.goalpanzi.domain.mission.MissionStatus.*;
 
 @Transactional(readOnly = true)
 @RequiredArgsConstructor
@@ -45,11 +41,9 @@ public class MissionMemberService {
     private final MissionRepository missionRepository;
     private final MemberRepository memberRepository;
     private final DeviceRepository deviceRepository;
-    private final DeviceSubscriptionRepository deviceSubscriptionRepository;
 
     private final ApplicationEventPublisher eventPublisher;
     private final PushMessageSender pushMessageSender;
-    private final TopicSubscriber topicSubscriber;
 
     public MissionDetailResponse getJoinableMission(final InvitationCode invitationCode) {
         missionValidator.validateJoinableMission(invitationCode);
@@ -155,36 +149,6 @@ public class MissionMemberService {
     }
 
     @Transactional
-    public void unsubscribeFromUselessMissions() {
-        List<Mission> missions = missionRepository.getInProgressMissions();
-
-        missions.forEach(mission -> {
-            List<MissionMember> missionMembers = missionMemberRepository.findAllWithMemberByMissionId(mission.getId());
-
-            if (isCancelledMission(missionMembers) || isCompletedMission(missionMembers)) {
-                eventPublisher.publishEvent(
-                        new UnsubscribeFromMissionEvent(mission.getId())
-                );
-                if (isCompletedMission(missionMembers)) {
-                    missionMembers.forEach(missionMember ->
-                            reserveRetryPushMessageForMember(missionMember.getMember())
-                    );
-                }
-            }
-        });
-    }
-
-    private boolean isCancelledMission(final List<MissionMember> missionMembers) {
-        return missionMembers.stream()
-                .anyMatch(it -> it.getMissionStatus().equals(CANCELED));
-    }
-
-    private boolean isCompletedMission(final List<MissionMember> missionMembers) {
-        return missionMembers.stream()
-                .anyMatch(it -> it.getMissionStatus().equals(COMPLETED));
-    }
-
-    @Transactional
     public void viewMissionRank(final Long missionId, final Long memberId) {
         MissionMember missionMember = missionMemberRepository.getMissionMember(memberId, missionId);
         missionMember.checkCompleted();
@@ -222,76 +186,5 @@ public class MissionMemberService {
                 );
             }
         });
-    }
-
-    @Transactional
-    public void subscribeToMyMissions(final Long memberId, final Long deviceId) {
-        Device device = deviceRepository.getDevice(deviceId);
-        List<String> topics = findMySubscribedTopics(deviceId);
-        List<Mission> missions = missionRepository.findAllById(
-                findMySubscribableMission(memberId, topics)
-        );
-
-        topics.forEach(topic ->
-                topicSubscriber.subscribeToTopic(List.of(device.getDeviceToken()), topic)
-        );
-        missions.forEach(mission -> {
-            deviceSubscriptionRepository.save(new DeviceSubscription(device, mission));
-
-            String topic = TopicGenerator.getTopic(mission.getId());
-            topicSubscriber.subscribeToTopic(List.of(device.getDeviceToken()), topic);
-        });
-        eventPublisher.publishEvent(
-                new UpdateMissionRetryPushMessageEvent(memberId, device.getDeviceToken())
-        );
-    }
-
-    public void unsubscribeFromMyMissions(final Long memberId, final Long deviceId, final String deprecatedDeviceToken) {
-        List<String> topics = findMySubscribedTopics(deviceId);
-
-        topics.forEach(topic ->
-                topicSubscriber.unsubscribeFromTopic(List.of(deprecatedDeviceToken), topic)
-        );
-        eventPublisher.publishEvent(
-                new CancelMissionRetryPushMessageEvent(memberId)
-        );
-    }
-
-    private List<String> findMySubscribedTopics(final Long deviceId) {
-        List<DeviceSubscription> subscriptions = deviceSubscriptionRepository.findAllWithMissionAndDeviceByDeviceId(deviceId);
-
-        return subscriptions.stream()
-                .map(it -> TopicGenerator.getTopic(it.getMission().getId()))
-                .toList();
-    }
-
-    private List<Long> findMySubscribableMission(final Long memberId, List<String> topicFilter) {
-        List<MissionStatus> statusFilter = List.of(CREATED, IN_PROGRESS, PENDING_COMPLETION);
-        List<MissionMember> missionMembers = missionMemberRepository.findAllWithMissionByMemberId(memberId);
-        List<MissionMember> filteredMissionMembers = missionMembers.stream()
-                .filter(it -> isMissionStatusMatching(statusFilter, it))
-                .filter(it -> isAlreadySubscribedMission(topicFilter, it.getMission().getId()))
-                .toList();
-
-        return filteredMissionMembers.stream()
-                .map(it -> it.getMission().getId())
-                .collect(Collectors.toList());
-    }
-
-    private boolean isAlreadySubscribedMission(List<String> filter, final Long missionId) {
-        return filter.contains(TopicGenerator.getTopic(missionId));
-    }
-
-    private void reserveRetryPushMessageForMember(final Member member) {
-        Devices devices = new Devices(
-                deviceRepository.findAllByMemberId(member.getId())
-        );
-
-        devices.getActivatedDeviceTokens()
-                .forEach(deviceToken ->
-                        eventPublisher.publishEvent(
-                                new ReserveMissionRetryPushMessageEvent(member.getId(), deviceToken)
-                        )
-                );
     }
 }
