+++ src/main/java/com/nexters/goalpanzi/application/mission/event/handler/MissionMemberEventHandler.java
@@ -2,12 +2,10 @@ package com.nexters.goalpanzi.application.mission.event.handler;
 
 import com.nexters.goalpanzi.application.firebase.TopicGenerator;
 import com.nexters.goalpanzi.application.member.event.DeleteMemberEvent;
-import com.nexters.goalpanzi.application.member.event.UpdateDeviceTokenEvent;
-import com.nexters.goalpanzi.application.member.event.UpdatePushActivationStatusEvent;
 import com.nexters.goalpanzi.application.mission.MissionMemberService;
+import com.nexters.goalpanzi.application.mission.MissionRetryPushMessageService;
 import com.nexters.goalpanzi.application.mission.MissionVerificationService;
-import com.nexters.goalpanzi.application.mission.event.CreateMissionEvent;
-import com.nexters.goalpanzi.application.mission.event.DeleteMissionEvent;
+import com.nexters.goalpanzi.application.mission.event.*;
 import com.nexters.goalpanzi.domain.mission.InvitationCode;
 import com.nexters.goalpanzi.infrastructure.firebase.PushMessageSender;
 import lombok.RequiredArgsConstructor;
@@ -28,6 +26,7 @@ public class MissionMemberEventHandler {
 
     private final MissionMemberService missionMemberService;
     private final MissionVerificationService missionVerificationService;
+    private final MissionRetryPushMessageService missionRetryPushMessageService;
 
     private final PushMessageSender pushMessageSender;
 
@@ -63,24 +62,23 @@ public class MissionMemberEventHandler {
     }
 
     @Async
-    @Transactional(propagation = Propagation.REQUIRES_NEW)
     @TransactionalEventListener(phase = TransactionPhase.AFTER_COMMIT)
-    void handleUpdateDeviceTokenEvent(final UpdateDeviceTokenEvent event) {
-        if (event.deprecatedDeviceToken() != null) {
-            missionMemberService.unsubscribeFromMyMissions(event.memberId(), event.deviceId(), event.deprecatedDeviceToken());
-        }
-        missionMemberService.subscribeToMyMissions(event.memberId(), event.deviceId());
-        log.info("Handled UpdateDeviceTokenEvent for memberId: {}", event.memberId());
+    void handleReserveMissionRetryPushMessageEvent(final ReserveMissionRetryPushMessageEvent event) {
+        missionRetryPushMessageService.reserveRetryPushMessage(event.memberId(), event.deviceToken());
+        log.info("Handled ReserveMissionRetryPushMessageEvent for memberId: {}", event.memberId());
     }
 
     @Async
     @TransactionalEventListener(phase = TransactionPhase.AFTER_COMMIT)
-    void handleUpdatePushActivationStatusEvent(final UpdatePushActivationStatusEvent event) {
-        if (event.isPushActivated()) {
-            missionMemberService.subscribeToMyMissions(event.memberId(), event.deviceId());
-        } else {
-            missionMemberService.unsubscribeFromMyMissions(event.memberId(), event.deviceId(), event.deviceToken());
-        }
-        log.info("Handled UpdatePushActivationStatusEvent for memberId: {}", event.memberId());
+    void handleCancelMissionRetryPushMessageEvent(final CancelMissionRetryPushMessageEvent event) {
+        missionRetryPushMessageService.cancelRetryPushMessage(event.memberId());
+        log.info("Handled CancelMissionRetryPushMessageEvent for memberId: {}", event.memberId());
+    }
+
+    @Async
+    @TransactionalEventListener(phase = TransactionPhase.AFTER_COMMIT)
+    void handleUpdateMissionRetryPushMessageEvent(final UpdateMissionRetryPushMessageEvent event) {
+        missionRetryPushMessageService.updateRetryPushMessage(event.memberId(), event.deviceToken());
+        log.info("Handled UpdateMissionRetryPushMessageEvent for memberId: {}", event.memberId());
     }
 }
